chore: move generic accordion component up
